+++ cli/npm/managed/resolvers/common/bin_entries.rs
